refactor: move to `utils/cloudflare.ts'
