+++ src/drivers/utils/index.ts
@@ -1,4 +1,3 @@
-/// <reference types="@cloudflare/workers-types" />
 import type { Driver } from "../../types";
 
 type DriverFactory<T> = (opts: T) => Driver;
@@ -39,32 +38,4 @@ export function createRequiredError(driver: string, name: string | string[]) {
     );
   }
   return createError(driver, `Missing required option \`${name}\`.`);
-}
-
-export function getBinding(binding: KVNamespace | R2Bucket | string) {
-  let bindingName = "[binding]";
-
-  if (typeof binding === "string") {
-    bindingName = binding;
-    binding = ((globalThis as any)[bindingName] ||
-      (globalThis as any).__env__?.[bindingName]) as KVNamespace | R2Bucket;
-  }
-
-  if (!binding) {
-    throw createError(
-      'Cloudflare',
-      `Invalid binding \`${bindingName}\`: \`${binding}\``
-    );
-  }
-
-  for (const key of ["get", "put", "delete"]) {
-    if (!(key in binding)) {
-      throw createError(
-        'Cloudflare',
-        `Invalid binding \`${bindingName}\`: \`${key}\` key is missing`
-      );
-    }
-  }
-
-  return binding;
 }
